refactor: 프로젝트 리스트 페이지 폴더 이름 변경
