+++ frontend/src/pages/project-list/CardList/Card/Card.styled.ts
