+++ frontend/src/pages/project-list/CardList/Card/StatBox/StatBox.styled.ts
