+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/CategoryFilterBox/CategoryFilterList/CategoryFilterButton/CategoryFilterButton.styled.ts
