+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/CategoryFilterBox/CategoryFilterList/CategoryFilterList.styled.ts
