+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/FilterBox/FilterBox.styled.ts
