+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/TechStackFilterBox/SelectedTechStack/SelectedTechStack.tsx
