+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/TechStackFilterBox/TechStackFilterList/TechStackFilterList.styled.ts
