+++ frontend/src/pages/project-list/FilterContainer/FilterTrigger/TechStackFilterBox/TechStackSearchBar/TechStackSearchBar.styled.ts
